+++ cleanup.sh
@@ -1,2 +1,4 @@
 #!/bin/bash
+cd "$(dirname "$0")" || exit 1
+
 /bin/rm -rf cache outputs .venv