+++ docker/build-image.sh
@@ -18,7 +18,7 @@ OPTS="--progress=plain --no-cache=true"
 #OPTS="--build-arg http_proxy=$http_proxy --build-arg https_proxy=$https_proxy"
 
 echo "===> build $target"
-docker build -f Dockerfile.$target -t tmurakam/kubespray-offline-$target:latest $OPTS .. || exit 1
+cd .. && sudo docker build -f docker/Dockerfile.$target -t tmurakam/kubespray-offline-$target:latest $OPTS . || exit 1
 
 if $push; then
     echo "===> push $target"