+++ docker/download-all.sh
@@ -3,4 +3,7 @@
 cd $(dirname $0)
 source ./common.sh
 
-run_in_docker ./download-all.sh
+run_in_docker "SKIP_DOWNLOAD_IMAGES=true ./download-all.sh"
+
+sudo chown -R $(id -u):$(id -g) ../outputs ../cache
+