+++ docker/download-files.sh
@@ -4,3 +4,6 @@ cd $(dirname $0)
 source ./common.sh
 
 run_in_docker "./prepare-py.sh && ./pypi-mirror.sh && SKIP_DOWNLOAD_IMAGES=true ./download-kubespray-files.sh"
+
+sudo chown -R $(id -u):$(id -g) ../outputs ../cache
+